Extracted chain downloading code to download.ts
